+++ src/bos/server/controllers/v2/sessiontemplates.py
@@ -145,8 +145,8 @@ def get_v2_sessiontemplate(session_template_id):
     if template_key not in DB:
         LOGGER.warning("Session template not found: %s", session_template_id)
         return connexion.problem(
-            status=404, title="Sessiontemplate could not found.",
-            detail="Sessiontemplate {} could not be found".format(session_template_id))
+            status=404, title="Session template not found.",
+            detail="Session template {} could not be found".format(session_template_id))
     template = DB.get(template_key)
     return template, 200
 
@@ -174,8 +174,8 @@ def delete_v2_sessiontemplate(session_template_id):
     if template_key not in DB:
         LOGGER.warning("Session template not found: %s", session_template_id)
         return connexion.problem(
-            status=404, title="Sessiontemplate could not found.",
-            detail="Sessiontemplate {} could not be found".format(session_template_id))
+            status=404, title="Session template not found.",
+            detail="Session template {} could not be found".format(session_template_id))
     return DB.delete(template_key), 204
 
 
@@ -191,8 +191,8 @@ def patch_v2_sessiontemplate(session_template_id):
     if template_key not in DB:
         LOGGER.warning("Session template not found: %s", session_template_id)
         return connexion.problem(
-            status=404, title="Sessiontemplate could not found.",
-            detail="Sessiontemplate {} could not be found".format(session_template_id))
+            status=404, title="Session template not found.",
+            detail="Session template {} could not be found".format(session_template_id))
 
     if connexion.request.is_json:
         LOGGER.debug("connexion.request.is_json")